move create duplicate mandataris to component
