+++ app/templates/organen/orgaan/mandatarissen.hbs
@@ -66,33 +66,7 @@
   </div>
 </AuModal>
 
-<AuModal
-  @title={{this.doubleMandateTitle}}
-  @modalOpen={{this.doubleMandateModal}}
-  @closable={{true}}
-  @closeModal={{this.toggleDoubleMandateModal}}
-  as |Modal|
->
-  <Modal.Body>
-    <p>
-      {{this.doubleMandateText}}
-    </p>
-  </Modal.Body>
-  <Modal.Footer>
-    <AuToolbar as |Group|>
-      <Group>
-        <AuButtonGroup>
-          <AuButton {{on "click" this.createDoubleMandataris}}>
-            Ja
-          </AuButton>
-          <AuButton
-            @skin="secondary"
-            {{on "click" this.toggleDoubleMandateModal}}
-          >
-            Nee
-          </AuButton>
-        </AuButtonGroup>
-      </Group>
-    </AuToolbar>
-  </Modal.Footer>
-</AuModal>
+<Mandatarissen::Linked::UpdateModal
+  @recentUpdate={{this.createdMandataris}}
+  @mandataris={{this.newestMandatarisId}}
+/>
